+++ spec/rley/parser/parse_tree_factory_spec.rb
@@ -1,40 +1,28 @@
 require_relative '../../spec_helper'
 
 require_relative '../../../lib/rley/parser/gfg_earley_parser'
-
 require_relative '../../../lib/rley/syntax/grammar_builder'
 require_relative '../support/grammar_helper'
+require_relative '../support/grammar_abc_helper'
 require_relative '../support/expectation_helper'
 
 # Load the class under test
-require_relative '../../../lib/rley/parser/parse_forest_factory'
+require_relative '../../../lib/rley/parser/parse_tree_factory'
 
 module Rley # Open this namespace to avoid module qualifier prefixes
   module Parser
-    describe ParseForestFactory do
+    describe ParseTreeFactory do
       include GrammarHelper     # Mix-in with token factory method
       include ExpectationHelper # Mix-in with expectation on parse entry sets
+      include GrammarABCHelper  # Mix-in for a sample grammar
 
       let(:sample_grammar) do
-          # Grammar based on paper from Elisabeth Scott
-          # "SPPF-Style Parsing From Earley Recognizers" in
-          # Notes in Theoretical Computer Science 203, (2008), pp. 53-67
-          # contains a hidden left recursion and a cycle
-          builder = Syntax::GrammarBuilder.new do
-            add_terminals('a', 'b')
-            rule 'Phi' => 'S'
-            rule 'S' => %w[A T]
-            rule 'S' => %w[a T]
-            rule 'A' => 'a'
-            rule 'A' => %w[B A]
-            rule 'B' => []
-            rule 'T' => %w[b b b]
-          end
-          builder.grammar
+        builder = grammar_abc_builder
+        builder.grammar
       end
 
       let(:sample_tokens) do
-        build_token_sequence(%w[a b b b], sample_grammar)
+        build_token_sequence(%w[a b c], sample_grammar)
       end
 
       let(:sample_result) do
@@ -44,7 +32,7 @@ module Rley # Open this namespace to avoid module qualifier prefixes
 
 
       subject do
-        ParseForestFactory.new(sample_result)
+        ParseTreeFactory.new(sample_result)
       end
 
       # Emit a text representation of the current path.
@@ -58,7 +46,7 @@ module Rley # Open this namespace to avoid module qualifier prefixes
 
       context 'Initialization:' do
         it 'should be created with a GFGParsing' do
-          expect { ParseForestFactory.new(sample_result) }.not_to raise_error
+          expect { ParseTreeFactory.new(sample_result) }.not_to raise_error
         end
 
         it 'should know the parse result' do
@@ -66,17 +54,10 @@ module Rley # Open this namespace to avoid module qualifier prefixes
         end
       end
 
-      context 'Parse forest construction' do
-        it 'should build a parse forest' do
-          forest = subject.build_parse_forest
-          expect(forest).to be_kind_of(SPPF::ParseForest)
-=begin
-          require 'yaml'
-          require_relative '../../../exp/lab/forest_representation'
-          File.open("forest.yml", "w") { |f| YAML.dump(forest, f) }
-          pen = ForestRepresentation.new
-          pen.generate_graph(forest, File.open("forest.dot", "w"))
-=end
+      context 'Parse tree construction' do
+        it 'should build a parse tree' do
+          forest = subject.create
+          expect(forest).to be_kind_of(PTree::ParseTree)
         end
       end # context
     end # describe